requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.6
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
uritemplate==4.1.1
